Removed regen from ActivateUnit() and moved Bulwark skill to Cleric directory
